Integrate madns into this project
This is what it's all been building up to! Because the madns server returns responses with fixed TTLs, and can send back "weird stuff" like fixed NS or PTR records that wouldn't be configurable in a normal resolver, we can test the _exact_ output of dog and assert that nothing has changed in the command-line interface. I'm really pleased about this!