Fix JSON output and introduce RecordType type
This commit fixes some bugs in the JSON output:

• The query type in the question structure is now the type name, like A, instead of a number like 1
• The record-specific data (like an A record's IP address) has been moved into a "data" field, for easier parsing
• The NAPTR regex field was incorrectly the service field
• The duration now has its structure specified, rather than relying on the internal representation
• Unknown classes now use their number, rather than a string of Rust Option

In order to turn types into names consistently, a `RecordType` type has been introduced to replace `TypeInt`. It's basically a `Record` without the assocatiated data.

Previously, even though a Query held a Labels field for a domain and a QClass field for a class, it still just held a u16 field for the record type. This was why it was showing up in the JSON output. It also showed up in the log output. Now, thanks to RecordType, its true name is known in both situations.

It also means we can get rid of some "weird stuff" like the qtype macro or the enum imports.

It also also means that, because RecordType is an enum, it will be a compile error if a new record type is added without the corresponding code in the wire module handling the new case.